Move renderBlog
