+++ apps/trpfrog.net/src/app/blog/_renderer/DevBlogMarkdown/hooks/useDevServerRenderedBlog.tsx
@@ -3,7 +3,7 @@ import { ReactNode, use, useDeferredValue, useEffect, useState } from 'react'
 import { createClient as createMdWatchClient } from '@trpfrog.net/dev-blog-server'
 
 import { ArticleSkeleton } from '@blog/[slug]/_components/ArticleSkeleton'
-import { renderBlog } from '@blog/_renderer/renderBlog'
+import { renderBlog } from '@blog/_renderer/DevBlogMarkdown/actions/renderBlog'
 
 const initializing = Symbol('trpfrog.net/blog-renderer:initializing')
